fix(B10): 修复 CPU 配额单位和配置验证
- 修复 H1: CPU 配额改为 float64 核心数，计算改为 * 1e9

- 修复 H2: 配置验证添加上下限检查

- 修复 H3: CreateContainer 添加运行时参数校验

+++ code/backend/internal/model/container.go
@@ -12,9 +12,9 @@ type ContainerConfig struct {
 
 // ResourceLimits 资源限制
 type ResourceLimits struct {
-	CPUQuota  int64
-	Memory    int64
-	PidsLimit int64
+	CPUQuota  float64 // CPU 核心数，如 0.5 表示 0.5 核
+	Memory    int64   // 内存限制（字节）
+	PidsLimit int64   // 进程数限制
 }
 
 // SecurityConfig 安全配置
